move string encoding to its own file
