+++ modules/c++/str/include/import/str.h
@@ -43,6 +43,7 @@
 #include "str/Tokenizer.h"
 #include "str/Format.h"
 #include "str/Manip.h"
+#include "str/Encoding.h"
 #define STR_MAJOR_VERSION 0
 #define STR_MINOR_VERSION 1
 #define STR_MICRO_VERSION 0
