+++ modules/c++/str/include/str/Convert.h
@@ -39,6 +39,8 @@
 // This is a fairly low-level file, so don't #include a lot of our other files
 #include "str/String_.h"
 
+#include "str/Encoding.h"
+
 namespace str
 {
 template <typename T>
@@ -73,125 +75,6 @@ std::string toString(const T& real, const T& imag)
 {
     return toString(std::complex<T>(real, imag));
 }
-
-template <typename TReturn, typename TChar>
-inline TReturn cast(const TChar* s)
-{
-    // This is OK as UTF-8 can be stored in std::string
-    // Note that casting between the string types will CRASH on some
-    // implementatons. NO: reinterpret_cast<const std::string&>(value)
-    const void* const pStr = s;
-    auto const retval = static_cast<TReturn>(pStr);
-    static_assert(sizeof(*retval) == sizeof(*s), "sizeof(*TReturn) != sizeof(*TChar)"); 
-    return retval;
-}
-template <typename TReturn, typename TChar>
-inline TReturn c_str(const std::basic_string<TChar>& s)
-{
-    return cast<TReturn>(s.c_str());
-}
-
-template <>
-inline std::string toString(const sys::U8string& value)
-{
-    return c_str<std::string::const_pointer>(value);  // copy
-}
-
-// This is to make it difficult to get encodings mixed up; it's here (in a .h
-// file) as we want to unit-test it. Windows1252_T for Windows-1252 characters
-enum class Windows1252_T : unsigned char { };  // https://en.cppreference.com/w/cpp/language/types
-using W1252string = std::basic_string<Windows1252_T>;  // https://en.cppreference.com/w/cpp/string
-template <>
-inline std::string toString(const W1252string& value)
-{
-    return c_str<std::string::const_pointer>(value);  // copy
-}
-
-void windows1252to8(W1252string::const_pointer, size_t, sys::U8string&); // c.f. utf16to8
-
-void utf16to1252(std::u16string::const_pointer, size_t, W1252string&);
-void utf32to1252(std::u32string::const_pointer, size_t, W1252string&);
-void wsto1252(std::wstring::const_pointer, size_t, W1252string&);
-
-// assume std::string is Windows-1252 **ON ALL PLATFORMS**
-sys::U8string fromWindows1252(std::string::const_pointer, size_t);
-inline sys::U8string fromWindows1252(const std::string& s)
-{
-    return fromWindows1252(s.c_str(), s.size());
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////
-// These use utf8:: routines; see utf8.h
-void utf16to8(std::u16string::const_pointer, size_t, sys::U8string&);
-void utf32to8(std::u32string::const_pointer, size_t, sys::U8string&);
-void wsto8(std::wstring::const_pointer, size_t, sys::U8string&);
-inline void strto8(std::u16string::const_pointer p, size_t sz, sys::U8string& result)
-{
-    utf16to8(p, sz, result);
-}
-inline void strto8(std::u32string::const_pointer p, size_t sz, sys::U8string& result)
-{
-    utf32to8(p, sz, result);
-}
-inline void strto8(std::wstring::const_pointer p, size_t sz, sys::U8string& result)
-{
-    wsto8(p, sz, result);
-}
-
-inline void utf16to8(const std::u16string& s, sys::U8string& result)
-{
-    utf16to8(s.c_str(), s.size(), result);
-}
-inline void utf32to8(const std::u32string& s, sys::U8string& result)
-{
-    utf32to8(s.c_str(), s.size(), result);
-}
-inline void wsto8(const std::wstring& s, sys::U8string& result)
-{
-    wsto8(s.c_str(), s.size(), result);
-}
-inline void strto8(const std::u16string& s, sys::U8string& result)
-{
-    utf16to8(s, result);
-}
-inline void strto8(const std::u32string& s, sys::U8string& result)
-{
-    utf32to8(s, result);
-}
-inline void strto8(const std::wstring& s, sys::U8string& result)
-{
-    wsto8(s, result);
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////
-
-bool mbsrtowcs(sys::U8string::const_pointer, size_t, std::wstring&);
-bool mbsrtowcs(const sys::U8string&, std::wstring&);
-bool wcsrtombs(std::wstring::const_pointer, size_t, sys::U8string&);
-bool wcsrtombs(const std::wstring&, sys::U8string&);
-
-// When the encoding is important, we want to "traffic" in either std::wstring (UTF-16 or UTF-32) or sys::U8string (UTF-8),
-// not str::W1252string (Windows-1252) or std::string (unknown).  Make it easy to get those from other encodings.
-std::wstring to_wstring(std::string::const_pointer, size_t); // assume Windows-1252 or UTF-8  based on platform
-std::wstring to_wstring(sys::U8string::const_pointer, size_t);
-std::wstring to_wstring(str::W1252string::const_pointer, size_t);
-template<typename TChar>
-inline std::wstring to_wstring(const std::basic_string<TChar>& s)
-{
-    return to_wstring(s.c_str(), s.size());
-}
-
-sys::U8string to_u8string(std::string::const_pointer, size_t);  // assume Windows-1252 or UTF-8  based on platform
-sys::U8string to_u8string(std::wstring::const_pointer, size_t);
-sys::U8string to_u8string(str::W1252string::const_pointer, size_t);
-template <typename TChar>
-inline sys::U8string to_u8string(const std::basic_string<TChar>& s)
-{
-    return to_u8string(s.c_str(), s.size());
-}
-
-// Encoding information is lost, will be UTF-8 or Windows-1252 or UTF-8 based on platform
-std::string to_string(const std::wstring&);
 
 template <typename T>
 T toType(const std::string& s)
@@ -344,16 +227,6 @@ T generic_cast(const std::string& value)
 {
     return str::toType<T>(value);
 }
-
-// Some string conversion routines only work with the right locale; the default
-// is platform-dependent.
-class setlocale final
-{
-    char* const locale_;
-public:
-    setlocale(const char* locale = "en_US.utf8");
-    ~setlocale() noexcept(false);
-};
 
 }
